E3DiffusionProcessのdiffuse_zero_to_t関数でノイズのdtypeをfloat32に変更。EquivariantGraphNeuralNetworkでMLPに隠れ層を追加。reverse_diffuse_one_step関数でalpha_sの計算をt-1に修正。test.pyで引数解析にnum_generateを追加し、データセットの読み込み処理を整理。train.pyでwandbのデータセットパスの更新を修正。

+++ E3diffusion_new.py
@@ -54,7 +54,7 @@ class E3DiffusionProcess():
         """
 
     def diffuse_zero_to_t(self,pos:torch.tensor,t:int):
-        noise = torch.zeros_like(pos)
+        noise = torch.zeros_like(pos,dtype=torch.float32)
         noise.normal_(mean=0,std=1)
         noise = remove_mean(noise)
         pos_after_diffuse = self.alpha(t) * pos + self.sigma(t) * noise
